% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgc_helpers.R
\name{download_cgc}
\alias{download_cgc}
\title{Download CGC files related to a particular update}
\usage{
download_cgc(url, destination)
}
\arguments{
\item{url}{a (character) string containing a url corresponding to the data page
for a CGC update or review.}

\item{download}{a (character) string corresponding to destination file in ./data-raw}
}
\description{
Download CGC files related to a particular update
}
