% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{estimated_gsp}
\alias{estimated_gsp}
\title{Gross State Product}
\format{
\subsection{\code{abs}}{

A data frame with nrow(revenue_assessed) rows and ncol(revenue_assessed) columns:
\describe{
\item{financial_year}{Financial year the estimate is for.}
\item{state_name}{The Australian State or Territory the estimate is for.}
\item{measure}{The type of value measured.}
\item{value}{The estimated value.}
...
}
}
}
\source{
\url{https://www.abs.gov.au/statistics/economy/government/taxation-revenue-australia}
}
\usage{
estimated_gsp
}
\description{
A dataset of the ABS's estimates of of state GSP.
}
\keyword{datasets}
