% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{expenses_gfs}
\alias{expenses_gfs}
\title{GFS expenses}
\format{
\subsection{\code{abs}}{

A data frame with nrow(revenue_assessed) rows and ncol(revenue_assessed) columns:
\describe{
\item{gfs_category}{The type of measure \code{expenses} to diffentiate from expenses.}
\item{gfs_subcategory}{The GFS \code{expenses} category.}
\item{state_name}{The Australian State or Territory the estimate is for.}
\item{financial_year}{Financial year the estimate is for.}
\item{value}{The estimated value.}
...
}
}
}
\source{
\url{https://www.abs.gov.au/statistics/economy/government/government-finance-statistics-annual/}
}
\usage{
expenses_gfs
}
\description{
A dataset of the ABS's estimates of state expenditure
}
\keyword{datasets}
