% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgc_helpers.R
\name{get_cgc_files}
\alias{get_cgc_files}
\title{Get CGC files related to a particular update}
\usage{
get_cgc_files(url)
}
\arguments{
\item{url}{a (character) string containing a url corresponding to the data page
for a CGC update or review.}
}
\description{
Get CGC files related to a particular update
}
