% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{latest_summary}
\alias{latest_summary}
\title{Latest relativity calculations}
\format{
\subsection{\code{cgc}}{

A data frame with nrow(revenue_assessed) rows and ncol(revenue_assessed) columns:
\describe{
\item{financial_year}{Financial year the estimate is for.}
\item{state_name}{The Australian State or Territory the estimate is for.}
\item{assessment_category}{The type of value assessed.}
\item{value}{The estimated value.}
...
}
}
}
\source{
\url{https://www.cgc.gov.au/reports-for-government/}
}
\usage{
latest_summary
}
\description{
A dataset of the CGC's calculations of relativities.
}
\keyword{datasets}
