% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{population_erp}
\alias{population_erp}
\title{Population}
\format{
\subsection{\code{abs}}{

A data frame with nrow(revenue_assessed) rows and ncol(revenue_assessed) columns:
\describe{
\item{financial_year}{Financial year the estimate is for.}
\item{state_name}{The Australian State or Territory the estimate is for.}
\item{unit}{The unit of population estimated.}
\item{population}{The estimated population.}
...
}
}
}
\source{
\url{https://www.abs.gov.au/statistics/people/population}
}
\usage{
population_erp
}
\description{
A dataset of the ABS's estimates of state population.
}
\keyword{datasets}
