% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{relativities_floorless}
\alias{relativities_floorless}
\title{Adjusted relativities}
\format{
\subsection{\code{cgc}}{

A data frame with nrow(relativities_floorless) rows and ncol(relativities_floorless) columns:
\describe{
\item{update_year}{Year of the CGC update or review the data was sourced from.}
\item{state_name}{The Australian State or Territory the relativity is for.}
\item{financial_year}{Financial year the relativity was to be used.}
\item{relativity}{The recommended relativity for that year as assessed by the CGC.}
\item{relativity_type}{The type of relativity \code{floorless} to diffentiate from recommended relativities.}
...
}
}
}
\source{
\url{https://www.cgc.gov.au/reports-for-government/}
}
\usage{
relativities_floorless
}
\description{
A dataset of relativities calculated on the basis a floor was never introduced.
}
\keyword{datasets}
