% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{revenue_adjusted}
\alias{revenue_adjusted}
\title{Adjusted revenue}
\format{
## `cgc`
A data frame with nrow(revenue_adjusted) rows and ncol(revenue_adjusted) columns:
\describe{
  \item{update_year}{Year of the CGC update or review the data was sourced from.}
  \item{state_name}{The Australian State or Territory the relativity is for.}
  \item{revenue_name}{The Australian State or Territory the relativity is for.}
  \item{financial_year}{Financial year the relativity was to be used.}
  \item{revenue}{The recommended relativity for that year as assessed by the CGC.}
  \item{revenue_type}{The type of estimate `fadjusted` to diffentiate from assessed and actual revenue.}
  ...
}
}
\source{
<https://www.cgc.gov.au/reports-for-government/>
}
\usage{
revenue_adjusted
}
\description{
A dataset of the CGC's estimates of state own source revenue.
}
\keyword{datasets}
