% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{revenue_assessed}
\alias{revenue_assessed}
\title{Assessed revenue}
\format{
## `cgc`
A data frame with nrow(revenue_assessed) rows and ncol(revenue_assessed) columns:
\describe{
  \item{update_year}{Year of the CGC update or review the data was sourced from.}
  \item{state_name}{The Australian State or Territory the relativity is for.}
  \item{revenue_name}{The Australian State or Territory the relativity is for.}
  \item{financial_year}{Financial year the relativity was to be used.}
  \item{revenue}{The recommended relativity for that year as assessed by the CGC.}
  \item{revenue_type}{The type of estimate `assessed` to diffentiate from adjusted and actual revenue.}
  ...
}
}
\source{
<https://www.cgc.gov.au/reports-for-government/>
}
\usage{
revenue_assessed
}
\description{
A dataset of the CGC's assessemtent of state own source revenue.
}
\keyword{datasets}
