% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{taxation_gfs}
\alias{taxation_gfs}
\title{GFS Taxation}
\format{
## `abs`
A data frame with nrow(revenue_assessed) rows and ncol(revenue_assessed) columns:
\describe{
  \item{gfs_category}{The type of measure `expenses` to diffentiate from expenses.}
  \item{gfs_subcategory}{The GFS `expense` category.}
  \item{state_name}{The Australian State or Territory the estimate is for.}
  \item{financial_year}{Financial year the estimate is for.}
  \item{value}{The estimated value.}
  ...
}
}
\source{
<https://www.abs.gov.au/statistics/economy/government/taxation-revenue-australia>
}
\usage{
taxation_gfs
}
\description{
A dataset of the ABS's estimates of of state taxation.
}
\keyword{datasets}
